requirements.txt
discord.py
Daug
python-dotenv
moviepy
